修正：由于 fork 过程可能导致共享 event_base 引发各种诡异问题；（考虑使用上层逻辑直接启用多进程）

+++ src/net/udp_socket.cpp
@@ -150,7 +150,7 @@ namespace net {
 					0, reinterpret_cast<struct sockaddr*>(&self->remote_addr_), &len);
 			}
 			php::callable cb = std::move(self->cb_);
-			if(self->rbuffer_->len < 0) {
+			if(ssize_t(self->rbuffer_->len) < 0) {
 				cb(core::make_exception(
 					boost::format("read failed: %s") % strerror(errno),
 					errno
